requirements.txt
discord.py==1.7.3
black
pylint
dateparser
git+git://github.com/jschaare/osrs-py.git